+++ src/hrdmc/estimators/__init__.py
@@ -5,7 +5,14 @@ from hrdmc.estimators.density import (
     estimate_open_line_density_profile,
     integrate_density_profile,
 )
-from hrdmc.estimators.local_energy import LocalEnergyResult, estimate_local_energy
+from hrdmc.estimators.local_energy import (
+    LocalEnergyResult,
+    TrappedLocalEnergyComponents,
+    TrappedLocalEnergyResult,
+    estimate_local_energy,
+    estimate_trapped_local_energy,
+    trapped_hard_rod_local_energy,
+)
 from hrdmc.estimators.pair_distribution import PairDistributionResult, estimate_pair_distribution
 from hrdmc.estimators.structure_factor import (
     StructureFactorResult,
@@ -14,6 +21,8 @@ from hrdmc.estimators.structure_factor import (
 
 __all__ = [
     "LocalEnergyResult",
+    "TrappedLocalEnergyComponents",
+    "TrappedLocalEnergyResult",
     "CloudMomentResult",
     "PairDistributionResult",
     "StructureFactorResult",
@@ -22,7 +31,9 @@ __all__ = [
     "estimate_density_profile",
     "estimate_open_line_density_profile",
     "estimate_local_energy",
+    "estimate_trapped_local_energy",
     "estimate_pair_distribution",
     "estimate_static_structure_factor",
     "integrate_density_profile",
+    "trapped_hard_rod_local_energy",
 ]